refactor pagination code for more general use
