+++ src/server/thingUtils.ts
@@ -1,6 +1,7 @@
 import type { APIContext } from "astro";
 import { jsonError, jsonResponse } from "./responses";
-import { and, db, eq, isDbError, Motion, Word, Sound, gt, asc, desc, lt } from "astro:db";
+import { and, db, eq, isDbError, Motion, Word, Sound } from "astro:db";
+import { paginationQuery } from "./pagination";
 
 const thingTypeToTable = {
     words: Word,
@@ -14,47 +15,23 @@ export function isThingType(value: string): value is ThingType {
 }
 
 export async function thingGet({ url }: APIContext, thingType: ThingType) {
+    const table = thingTypeToTable[thingType];
+
     const params = url.searchParams;
     const discordID = params.get("discord");
     if (!discordID || typeof discordID !== "string")
         return jsonError("Invalid Discord ID");
 
-    const direction = params.get("direction") ?? "next";
-    const cursor = parseInt(params.get("cursor") ?? (direction === "next" ? "0" : "999999999999"));
-    const pageSize = parseInt(params.get("pageSize") ?? "10");
-
-    if (isNaN(cursor) || isNaN(pageSize))
-        return jsonError("Invalid cursor or pageSize");
-    if (direction !== "next" && direction !== "prev")
-        return jsonError("Invalid direction");
-
     const showDeleted = params.get("showDeleted");
     if (showDeleted && showDeleted !== "true" && showDeleted !== "false")
         return jsonError("Invalid showDeleted");
 
-    const table = thingTypeToTable[thingType];
-
-    const things = await db
-        .select()
-        .from(table)
-        .where(
-            and(
-                direction === "next" ? gt(table.id, cursor) : lt(table.id, cursor),
-                eq(table.memberDiscord, discordID),
-                eq(table.deleted, showDeleted === "true"),
-            )
-        )
-        .limit(pageSize)
-        .orderBy(direction === "next" ? asc(table.id) : desc(table.id));
-
-    if (things.length && direction === "prev")
-        things.reverse();
-
-    return jsonResponse({
-        things,
-        prevCursor: things[0]?.id ?? undefined,
-        nextCursor: things.length === pageSize ? things[things.length - 1]?.id : undefined,
-    });
+    return paginationQuery(
+        params,
+        table,
+        eq(table.memberDiscord, discordID),
+        eq(table.deleted, showDeleted === "true"),
+    );
 }
 
 export async function thingDeletion({ url }: APIContext, thingType: ThingType, isDeleted: boolean) {
